Move store.ts into lib
