+++ src/channel-header.tsx
@@ -12,7 +12,7 @@ import { ChannelMembersViewModel, ChannelMembersView } from './channel-members-v
 import { isDM } from './lib/models/slack-api';
 import { Model, fromObservable } from './lib/model';
 import { SimpleView } from './lib/view';
-import { Store } from './store';
+import { Store } from './lib/store';
 import { when } from './lib/when';
 import { Observable } from 'rxjs/Observable';
