+++ src/channel-list.tsx
@@ -6,7 +6,7 @@ import { ChannelViewModel, ChannelListItem } from './channel-list-item';
 import { channelSort, isDM } from './lib/models/slack-api';
 import { CollectionView } from './lib/collection-view';
 import { fromObservable, notify, Model } from './lib/model';
-import { Store, ChannelList } from './store';
+import { Store, ChannelList } from './lib/store';
 import { when } from './lib/when';
 
 @notify('selectedChannel')
